+++ permissionmanager/src/main/java/io/jasonatwood/permissionmanager/DangerousPermissions.java
@@ -1,4 +1,4 @@
-package com.bignerdranch.permissionmanager;
+package io.jasonatwood.permissionmanager;
 
 
 import android.Manifest.permission;
