+++ permissionmanager/src/main/java/io/jasonatwood/permissionmanager/PermissionListener.java
@@ -1,4 +1,4 @@
-package com.bignerdranch.permissionmanager;
+package io.jasonatwood.permissionmanager;
 
 public interface PermissionListener {
     void onResult(boolean permissionGranted);
