+++ permissionmanager/src/main/java/io/jasonatwood/permissionmanager/PermissionManager.java
@@ -1,4 +1,4 @@
-package com.bignerdranch.permissionmanager;
+package io.jasonatwood.permissionmanager;
 
 import android.app.Activity;
 import android.content.Intent;
