+++ sample/src/main/java/io/jasonatwood/permissionmanagersample/MainActivity.java
@@ -1,10 +1,12 @@
-package com.bignerdranch.permissionmanagersample;
+package io.jasonatwood.permissionmanagersample;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+
+
 
 public class MainActivity extends AppCompatActivity {
