+++ sample/src/main/java/io/jasonatwood/permissionmanagersample/RequestPermissionFragment.java
@@ -1,4 +1,4 @@
-package com.bignerdranch.permissionmanagersample;
+package io.jasonatwood.permissionmanagersample;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
