+++ sample/src/main/java/io/jasonatwood/permissionmanagersample/RequestStoragePermissionFragment.java
@@ -1,17 +1,16 @@
-package com.bignerdranch.permissionmanagersample;
+package io.jasonatwood.permissionmanagersample;
 
 import android.Manifest;
 import android.support.v4.app.Fragment;
 
-import com.bignerdranch.permissionmanager.PermissionListener;
-import com.bignerdranch.permissionmanager.PermissionManager;
+import io.jasonatwood.permissionmanager.PermissionListener;
+import io.jasonatwood.permissionmanager.PermissionManager;
 
-public class RequestContactsPermissionFragment extends RequestPermissionFragment {
-
+public class RequestStoragePermissionFragment extends RequestPermissionFragment {
     private PermissionListener mListener;
 
     public static Fragment newInstance() {
-        return new RequestContactsPermissionFragment();
+        return new RequestStoragePermissionFragment();
     }
 
     @Override
@@ -31,11 +30,11 @@ public class RequestContactsPermissionFragment extends RequestPermissionFragment
         PermissionManager.askForPermission(getActivity(),
                 getPermission(),
                 mListener,
-                "Fragment needs Contacts permission to blah blah. This app won't work without them.");
+                "Fragment needs Storage permission to blah blah. This app won't work without them.");
     }
 
     @Override
     protected String getPermission() {
-        return Manifest.permission.GET_ACCOUNTS;
+        return Manifest.permission.WRITE_EXTERNAL_STORAGE;
     }
 }
