+++ sample/src/main/java/io/jasonatwood/permissionmanagersample/TwoComponentsSamePermissionActivity.java
@@ -1,4 +1,4 @@
-package com.bignerdranch.permissionmanagersample;
+package io.jasonatwood.permissionmanagersample;
 
 import android.content.Context;
 import android.content.Intent;
