+++ sample/src/main/java/io/jasonatwood/permissionmanagersample/TwoPaneActivity.java
@@ -1,4 +1,4 @@
-package com.bignerdranch.permissionmanagersample;
+package io.jasonatwood.permissionmanagersample;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
