+++ sample/src/main/res/layout/activity_main.xml
@@ -1,34 +1,37 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
-    android:id="@+id/fragment_container"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    android:padding="16dp"
-    tools:context="com.bignerdranch.permissionmanagersample.MainActivity">
+    android:padding="16dp">
 
-    <TextView android:layout_width="match_parent" android:layout_height="wrap_content"
-              android:text="We'll use an Activity hosting two Fragments in a two-pane layout to represent multiple components, each requesting permission(s) at the same time. \n\n In reality, *any combination* of Fragments and Activities can asynchronously and in parallel call into Permission Manager and request a permission."/>
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="We'll use an Activity hosting two Fragments in a two-pane layout to represent multiple components, each requesting permission(s) at the same time. \n\n In reality, *any combination* of Fragments and Activities can asynchronously and in parallel call into Permission Manager and request a permission."/>
 
-    <Button android:id="@+id/activity_main_single_permission_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:text="Single permission request"/>
+    <Button
+        android:id="@+id/activity_main_single_permission_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:text="Single permission request"/>
 
-    <Button android:id="@+id/activity_main_two_components_same_permission_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:text="Two components, both requesting the same permission."/>
+    <Button
+        android:id="@+id/activity_main_two_components_same_permission_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:text="Two components, both requesting the same permission."/>
 
-    <Button android:id="@+id/activity_main_two_components_different_permission_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:text="Two components, each requesting a different permission."/>
+    <Button
+        android:id="@+id/activity_main_two_components_different_permission_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:text="Two components, each requesting a different permission."/>
 
 
 </LinearLayout>
